% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverseRCT.r
\name{defaults_prep}
\alias{defaults_prep}
\title{Default preprocessing pipelines}
\usage{
defaults_prep(na_threshold = 0.2)
}
\arguments{
\item{na_threshold}{Maximum proportion of missing values allowed before
excluding a variable (used in the \code{drop_sparse} method)}
}
\value{
A named list of preprocessing functions
}
\description{
Default preprocessing pipelines
}
