% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverseRCT.r
\name{extract_models}
\alias{extract_models}
\title{Extract models from multiverse results}
\usage{
extract_models(x, prep_name = NULL, model_name = NULL)
}
\arguments{
\item{x}{A multiverse_rct object}

\item{prep_name}{Preprocessing method name(s) or NULL for all}

\item{model_name}{Model specification name(s) or NULL for all}
}
\value{
A list of fitted model objects
}
\description{
Get the fitted model objects from specific specifications.
}
