% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverseRCT.r
\name{filter_multiverse}
\alias{filter_multiverse}
\title{Filter multiverse results}
\usage{
filter_multiverse(
  x,
  min_effect = NULL,
  max_effect = NULL,
  min_n = NULL,
  prep = NULL,
  model = NULL,
  significant_only = FALSE
)
}
\arguments{
\item{x}{A multiverse_rct object}

\item{min_effect}{Minimum effect size to include}

\item{max_effect}{Maximum effect size to include}

\item{min_n}{Minimum sample size to include}

\item{prep}{Preprocessing method(s) to include}

\item{model}{Model specification(s) to include}

\item{significant_only}{If TRUE, only include models with p < 0.05}
}
\value{
A filtered multiverse_rct object
}
\description{
Select a subset of the multiverse based on various criteria.
}
