% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverseRCT.r
\name{mean_impute}
\alias{mean_impute}
\title{Mean imputation for numeric columns}
\usage{
mean_impute(df, cols = NULL)
}
\arguments{
\item{df}{Data frame}

\item{cols}{Character vector of columns or NULL for all numeric columns}
}
\description{
Mean imputation for numeric columns
}
