% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverseRCT.r
\name{median_impute}
\alias{median_impute}
\title{Median imputation for numeric columns}
\usage{
median_impute(df, cols = NULL)
}
\arguments{
\item{df}{Data frame}

\item{cols}{Character vector of columns or NULL for all numeric columns}
}
\description{
Median imputation for numeric columns
}
