% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverseRCT.r
\name{mode_impute}
\alias{mode_impute}
\title{Mode imputation for categorical columns}
\usage{
mode_impute(df, cols = NULL)
}
\arguments{
\item{df}{Data frame}

\item{cols}{Character vector of columns or NULL for all factor/character columns}
}
\description{
Mode imputation for categorical columns
}
