% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverseRCT.r
\name{remove_outliers}
\alias{remove_outliers}
\title{Outlier detection and removal}
\usage{
remove_outliers(
  df,
  cols = NULL,
  method = c("z_score", "iqr", "percentile"),
  threshold = 3
)
}
\arguments{
\item{df}{Data frame}

\item{cols}{Character vector of columns or NULL for all numeric columns}

\item{method}{Method for outlier detection: "z_score", "iqr", or "percentile"}

\item{threshold}{Threshold for outlier detection}
}
\description{
Outlier detection and removal
}
