% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverseRCT.r
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Sensitivity analysis for multiverse results}
\usage{
sensitivity_analysis(x)
}
\arguments{
\item{x}{A multiverse_rct object}
}
\value{
A list of sensitivity metrics
}
\description{
Assess the stability of findings across the analytic multiverse.
}
